Create a ggnetwork data frame from an igraph object

Create a ggnetwork data frame from an igraph object

igraph2ggnetwork(graph, layout = "kk", arrow.gap = 0.2)

Arguments

graph
Object of class igraph.

layout
Network layout. One of "dh", "drl", "gem", "lgl", "fr", "graphopt", "kk" and "mds". Default is "kk".

arrow.gap
Numeric indicating the distance between nodes and arrows. Default is 0.2

See also

Author

Fabricio Almeida-Silva