Plot expression profile of given genes across samples

Plot expression profile of given genes across samples

plot_expression_profile(
  genes,
  exp,
  metadata,
  plot_module = TRUE,
  genes_modules,
  modulename
)

Arguments

genes
Character vector containing a subset of genes from which edges will be extracted. It can be ignored if plot_module is TRUE.

exp
Data frame containing genes IDs in row names and sample names in column names.

metadata
A 2-column data frame containing sample names in the first column and sample descriptions in the second column.

plot_module
Logical indicating whether to plot a whole module or not. If set to FALSE, genes must be specified.

genes_modules
Data frame containing genes in column 1 and their corresponding modules in column 2. It is the third element of the output list from exp2net.

modulename
Character with name of the module to be plotted. To include 2 or more modules, input the names in a character vector.

Value

A ggplot object showing the expression profile of some genes across all samples.

See also

Author

Fabricio Almeida-Silva