Plot heatmap of hierarchically clustered sample correlations or gene expression

Plot heatmap of hierarchically clustered sample correlations or gene expression

plot_heatmap(
  exp,
  col_metadata = NULL,
  row_metadata = NULL,
  cor_method = "spearman",
  type = "samplecor",
  palette = NULL,
  log_trans = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  show_rownames = FALSE,
  show_colnames = TRUE,
  scale = "none",
  fontsize = 10,
  cutree_rows = 1,
  cutree_cols = 1
)

Arguments

exp
Expression data frame with genes in rows and samples in columns.

col_metadata
A data frame containing sample names in the first column and sample description in the subsequent columns. If there are more than 2 columns, additional columns will be interpreted as multiple column annotations.

row_metadata
A data frame containing gene IDs in the first column and gene functional classification in the subsequent columns. If there are more than 2 columns, additional columns will be interpreted as multiple gene annotation classifications.

cor_method
Correlation method. One of 'spearman' or 'pearson'. Default is 'spearman'.

type
Type of heatmap to plot. One of 'samplecor' (sample correlations) or 'expr'. Default is 'samplecor'.

palette
RColorBrewer palette to use. Default is "Blues" for sample correlation heatmap and "YlOrRd" for gene expression heatmap.

log_trans
Logical. It specifies whether to log transform data or not. Default is FALSE.

cluster_rows
Logical indicating whether to cluster rows or not. Default is TRUE.

cluster_cols
Logical indicating whether to cluster columns or not. Default is TRUE.

show_rownames
Logical indicating whether to show row names or not. Default is FALSE.

show_colnames
Logical indicating whether to show column names or not. Default is TRUE.

scale
Character indicating if values should be centered and scaled in rows, columns, or none. One of 'row', 'column', or 'none'. Default is 'none'.

fontsize
Base fontsize for the plot.

cutree_rows
Number of clusters into which rows are divided. Default is 1, which is equivalent to no division.

cutree_cols
Number of clusters into which columns are divided. Default is 1, which is equivalent to no division.

Value

heatmap of hierarchically clustered samples with metadata information (optional)

See also

Author

Fabricio Almeida-Silva