Plot network from edge list

Plot network from edge list

plot_network(
  bionetwork = "GCN",
  edgelist_gcn = NULL,
  edgelist_int = NULL,
  genes_modules = NULL,
  modulename = NULL,
  kIN = NULL,
  hubs = NULL,
  top_n_hubs = 5,
  hubs_only = FALSE,
  layout = layout_with_drl(),
  facet_mod = TRUE,
  clustering_method = "infomap"
)

Arguments

bionetwork
Type of biological network. One of 'GCN' (gene coexpression network), 'PPI' (protein-protein interaction), 'GRN' (gene-regulatory network), or 'Mixed' (GCN + PPI). Default is 'GCN'.

edgelist_gcn
Data frame containing the edge list for the gene coexpression network. First column is gene 1 and second column is gene 2, which is coexpressed with gene 1. All other columns are interpreted as edge attributes (e.g. weight). The edge list can be obtained from an adjacency matrix with the function get_edge_list. If bionetwork is 'PPI' or 'GRN', this parameter can be ignored.

edgelist_int
Data frame containing the edge list for the PPI or GRN network. First column is the protein 1 (or TF) and second column is the protein 2 (or target gene). All other columns are interpreted as edge attributes. If bionetwork is 'GCN', this parameter can be ignored.

genes_modules
Data frame containing genes in column 1 and their corresponding modules in column 2. It is the third element of the output list from exp2net.

modulename
Character with name of the module to be plotted. To include 2 or more modules, input the names in a character vector.

kIN
Data frame contaning intramodular connectivity for each gene in the network. It is the fourth element of the output list from exp2net.

hubs
Data frame containing hub genes in the first column, their modules in the second column, and intramodular connectivity in the third column.

top_n_hubs
Number of top hubs to be labeled. Default is 5.

hubs_only
Logical indicating if only hub genes should be plotted. Default is FALSE.

clustering_method
Community detection algorithm to be used. Available methods are "infomap", "edge_betweenness", "fast_greedy", "walktrap", "spinglass", "leading_eigen", "louvain", and "label_prop". Default is "infomap".

facet
Logical. If more than 1 module is chosen, should different modules be plotted in different facets? Default is TRUE.

Value

A ggplot object with the network.

See also