Identify consensus modules across multiple data sets

This function identifies modules that are conserved across multiple data sets. It can be used for evolutionary analyses and for testing robustness of coexpression modules in data sets from different sources, for instance.

consensus_modules(
  exp_list,
  setLabels = NULL,
  metadata,
  cor_method = "spearman",
  net_type = "signed hybrid",
  rsquared = 0.8,
  module_merging_threshold = 0.2
)

Arguments

exp_list
A list containing the expression data frames. For each expression data frame, row names correspond to gene IDs and column names correspond to sample names. The list can be created by using list(exp1, exp2, ..., expn).

setLabels
Character vector containing labels for each expression set.

metadata
Data frame containing sample information. The first column must contain sample names and the second column must contain sample information.

cor_method
Correlation method used for network reconstruction. One of "spearman" (default), "biweight", or "pearson".

net_type
Network type. One of "signed hybrid" (default), "signed" or "unsigned".

rsquared
Minimum R squared to consider the network similar to a scale-free topology. Default is 0.8.

module_merging_threshold
Cut-off for merging similar modules based on the dissimilarity of their eigengenes. Default is 0.2, indicating that modules that have correlation >= 0.8 will be merged into one.

Value

A list containing 4 elements:

consModules

Consensus module assignments

consMEs

Consensus module eigengenes

exprSize

Description of the multi-set object returned by the function WGCNA::checkSets

sampleInfo

Metadata for each expression set

See also