Perform enrichment analysis for a set of genes

Perform enrichment analysis for functional annotation such as Gene Ontology (GO), pathway (KEGG, MapMan) or protein domains (PFAM, Panther).

enrichment_analysis(
  genes,
  exp,
  annotation,
  column,
  correction = "BH",
  p = 0.05
)

Arguments

genes
Character vector containing genes for overrepresentation analysis.

exp
Gene expression data frames with gene IDs in row names and column names in column names.

annotation
Annotation data frame with genes in the first column and functional annotation in the other columns, which can be exported from Biomart or similar databases.

column
Column name on the data frame specified in 'annotation' used to test for enrichment (e.g. "GO.ID").

correction
Multiple testing correction method. One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none". Default is "BH".

p
P-value threshold. P-values below this threshold will be considered significant. Default is 0.05.

Value

List containing two elements:

Significant_terms

Data frame contaning significant terms and p-values

Associated_genes

List containing genes associated with each significant term

See also

Author

Fabricio Almeida-Silva