Reconstruct gene coexpression network from gene expression

Reconstruct gene coexpression network from gene expression

exp2gcn(
  exp,
  net_type = "signed",
  module_merging_threshold = 0.8,
  SFTpower = NULL,
  cor_method = "spearman",
  verbose = FALSE
)

Arguments

exp

A gene expression data frame with genes in row names and samples in column names or a `SummarizedExperiment` object.

net_type

Network type. One of 'signed', 'signed hybrid' or 'unsigned'. Default: 'signed'.

module_merging_threshold

Correlation threshold to merge similar modules into a single one. Default: 0.8.

SFTpower

SFT power generated by the function SFT_fit.

cor_method

Correlation method. One of "pearson", "biweight" or "spearman". Default is "spearman".

verbose

Logical indicating whether to display progress messages or not. Default: FALSE.

Value

List containing:

Adjacency matrix

Data frame of module eigengenes

Data frame of genes and their corresponding modules

Data frame of intramodular connectivity

Vector of module assignment

Correlation matrix

Parameters used for network reconstruction

Objects to plot the dendrogram in plot_dendro_and_colors.

See also

Author

Fabricio Almeida-Silva

Examples

data(filt.se)
# The SFT fit was previously calculated and the optimal power was 16
gcn <- exp2gcn(filt.se, SFTpower = 18, cor_method = "pearson")
#> ..connectivity..
#> ..matrix multiplication (system BLAS)..
#> ..normalization..
#> ..done.