Calculate gene significance for a given group of genes

Calculate gene significance for a given group of genes

gene_significance(
  exp,
  metadata,
  genes = NULL,
  alpha = 0.05,
  min_cor = 0.2,
  use_abs = TRUE,
  palette = "RdYlBu",
  show_rownames = FALSE,
  continuous_trait = FALSE
)

Arguments

exp

A gene expression data frame with genes in row names and samples in column names or a `SummarizedExperiment` object.

metadata

A data frame containing sample names in row names and sample annotation in the first column. Ignored if `exp` is a `SummarizedExperiment` object, since the function will extract colData.

genes

Character vector of genes to be correlated with traits. If not given, all genes in `exp` will be considered.

alpha

Significance level. Default is 0.05.

min_cor

Minimum correlation coefficient. Default is 0.2.

use_abs

Logical indicating whether to filter by correlation using absolute value or not. If TRUE, a min_cor of say 0.2 would keep all correlations above 0.2 and below -0.2. Default is TRUE.

palette

RColorBrewer's color palette to use. Default is "RdYlBu", a palette ranging from blue to red.

show_rownames

Logical indicating whether to show row names or not. Default is FALSE.

continuous_trait

Logical indicating if trait is a continuous variable. Default is FALSE.

Value

A heatmap of gene significance (GS) and a list containing:

filtered_corandpFiltered matrix of correlation and p-values

raw_GSRaw matrix of gene significances

See also

Examples

data(filt.se)
gs <- gene_significance(filt.se)