Plot gene regulatory network from edge list

Plot gene regulatory network from edge list

plot_grn(
  edgelist_grn,
  show_labels = "tophubs",
  top_n_hubs = 5,
  interactive = FALSE,
  layout = "kk",
  arrow.gap = 0.01
)

Arguments

edgelist_grn
Data frame containing the edge list for the GRN network. First column is the TF and second column is the target gene. All other columns are interpreted as edge attributes.

show_labels
Character indicating which nodes will be labeled. One of "all", "allhubs", "tophubs", or "none".

top_n_hubs
Number of top hubs to be labeled. It is only valid if show_labels equals "tophubs". Default is 5.

interactive
Logical indicating whether the network should be interactive or not. Default is FALSE.

layout
Network layout. One of "dh", "drl", "gem", "lgl", "fr", "graphopt", "kk" and "mds". Default is "kk".

arrow.gap
Numeric indicating the distance between nodes and arrows. Default is 0.2.

Value

A ggplot object containing the network.

See also

Author

Fabricio Almeida-Silva